fabric-rag/requirements.txt
open-clip-torch>=2.26.0
faiss-cpu>=1.8.0
pyyaml>=6.0.1
pydantic>=2.6.0
torch>=2.2.0
numpy>=1.26.0
pandas>=2.2.0
pyarrow>=15.0.0
rich>=13.7.0
